Adjust journal updates in generic ERC20 transfer
The balance changes made on the Cosmos level by
the given precompile should be propagated back
to the EVM stateDB (using the journal) IF AND ONLY
IF the given precompile is tied to the native gas
token of the EVM layer. This is absolutely CRITICAL to properly sync state changes between the Cosmos SDK and EVM layers.

On the other hand, this MUST NOT be done if the precompile is tied to a non-gas ERC20 token to ensure its transfers do not affect the balances of the EVM native gas token for the from/to addresses.