Fix security advisory GHSA-2hq6-q2xv-9jgv
In the addVote method of the assetLockedVoteCounter type, when comparing
assetLockVoteInfo events, only the amount and recipient were used. This
could have lead to a malicious validator to use an invalid vote with
an empty token address to vote, and result in funds being locked in the
bridge. We now added a comparison using the token address as well.

A new test have been added, the results of the tests can be compared
by removing the changes in the file x/bridge/abci/proposal.go and
running the specific test using the following command:
-
go test -v -run="TestAssetsLockedExtractorTestSuite/TestCanonicalEvents/\
GHSA-2hq6-q2xv-9jgv_-_ensure_the_token_address_is_included_in_the_comparison" \
./x/bridge/...
-

Without the change the test fail, an extra event is generated (the
wrongful assetlock event), with the changes, no event are generated,
and the decisions of the validators stay split between the 2 different
token using the same sequence number.

Signed-off-by: Jeremy Letang <me@jeremyletang.com>