Parametrize mainnet explorer RPC
Here we add a way to parametrize the RPC used by the mainnet block explorer. Moreover, we have it in a secret for additional privacy.

By the way, we are also lowering the value of `COIN_BALANCE_ON_DEMAND_FETCHER_THRESHOLD` which should result with faster balance updates, especially after the bridge action.