Fix behaviour of the ERC20.approve(0)
The ERC20 std implementation (e.g: openzeppelin) allow a user
to approve 0 with success. This is to allow such expected pattern
where a user would first call approve(0) then approve(actualAmount)
on the contract. Until now the Mezo precompile would actually return an
error an revert if the allowance was already 0.

This PR fixes this behaviour, and basically makes a call to approve(0)
a no op effectively.

Signed-off-by: Your Name <me@jeremyletang.com>