Fix signature r + s values concatenation in permit
Previously the permit method on the ERC20 precompile could eventually
concatenate a wrong signature, resolving to an error. Here we concatenate
the original bytes array into a slice instead of using the .Byte from
big.Int which would trimLeft on the bytes.

We also uncomment the whole permit test suite which was previously
commented specifically because of this issue.

Signed-off-by: Your Name <me@jeremyletang.com>