Collect metrics from the Portal API
The Portal API exposes health metrics that are useful for monitoring.
In this commit, we configure Prometheus to scrape the Portal API.

The Portal API is deployed to staging and production environments, and
the metrics are collected from both environments separately. I failed to
find a cleaner way to use kustomize templates to generate the configmap,
so I created a local copy of the configmap for each environment.

Reference PR for Portal API: https://github.com/thesis/mezo-portal/pull/2076