Re-generate bindings for Sepolia
We upgraded `MezoBridge` on Sepolia
so it now contains new features related to bridge out. We now need to update the bindings.

Our bindings use a simplified approach where Ethereum Mainnet bindings are the default for all Ethereum networks. This approach works until there are no differences between contracts deployed on Ethereum networks. In this particular case, Sepolia is ahead until we deplpy the same on Mainnet. Because of that, we temporarily make Sepolia bindings the default ones. We will cleanup this during https://linear.app/thesis-co/issue/TET-1197/mainnet-deployment-phase-i-bridge-out.