Update MezoBridge mainnet bindings
Some time ago, we deployed the 2-way bridging capability on Mezo Testnet. To make it work, we had to switch `mezod` bindings to the version of `MezoBridge` available on Sepolia (as the 2-way bridge features of `MezoBridge` were not yet available on Ethereum Mainnet). This was done as part of https://github.com/mezo-org/mezod/pull/532.

Now, we upgraded the `MezoBridge` on Ethereum Mainnet and we can pin back `mezod` bindings to this version.