Configurable addresses cap in the filter query for logs calls
So far, we used a global limit of 10 addresses that could be used in the filter query while calling `eth_getLogs` and `eth_subscribe` with `logs` parameter.  However, this turned out to be very problematic for Goldsky indexers that query logs from multiple contracts at once.
To workaround this problem, we are making this limit configurable through the `json-rpc.logs-filter-addr-cap` config parameter and flag. We are also raising the default limit to 30.